Module bittorrent::error

Structs

ParseError
Used when parsing external data that may have errors at any position within the buffer.

TorrError
Used to raise an error when a piece of data required by the Torrent is missing from the Bencode data.

Enums

TorrErrorKind
A list specifying the types of TorrErrors that may occur.

Type Definitions

ParseResult

TorrResult