Module bittorrent::peerUnstable

Facilitates communication with a remote peer.

Modules

data

handshake

state

Structs

Peer
A Peer object representing a connection to a remote peer as well as the bittorrent state associated with that connection. Since the messaging interface between a peer is asynchronous, peers will start out as not choked and not interested.

Enums

PeerAction
Denotes an action that was carried out in response to a message received by a remote peer.