Module bittorrent::peer::blockUnstable

Unit of data being sent from a remote peer.

Structs

Block
Represents a block of data within a piece. Many blocks make up a piece, many pieces make up a file, and many (or one) file(s) make up the data being pointed to by a torrent file.