Struct bittorrent::peer::block::BlockUnstable

pub struct Block {
    // some fields omitted
}

Represents a block of data within a piece. Many blocks make up a piece, many pieces make up a file, and many (or one) file(s) make up the data being pointed to by a torrent file.

Methods

impl Block

fn with_capacity(capacity: BlockLength) -> Block

Creates a block with capacity size.

fn is_active(&self) -> bool

Returns true if the current block is marked as active.

fn set_active(&mut self, index: PieceIndex, offset: BlockOffset, length: BlockLength)

Marks the current block as active and sets the block information passed in.

fn set_inactive(&mut self)

Marks the current block as inactive.

fn as_mut_slice(&mut self) -> &mut [u8]

Returns a mutable slice to the data of this block.

fn piece_index(&self) -> PieceIndex

Returns the zero based piece index for the piece that this block belongs to.

fn block_offset(&self) -> BlockOffset

Returns the byte offset within the piece that this block belongs to.

fn block_len(&self) -> BlockLength

Returns the block length for the current block.