Trait bittorrent::peer::data::DataSender

pub trait DataSender {
    fn send_request(&mut self, piece: u32, offset: u32, length: u32) -> IoResult<()>;
    fn send_block(&mut self, piece: u32, offset: u32, block: &[u8]) -> IoResult<()>;
    fn send_cancel(&mut self, piece: u32, offset: u32, length: u32) -> IoResult<()>;
}

A trait dealing with data based messaging between a local and remote peer. This should be used as an interface for implementing the data messaging interface of the Bittorrent Peer Wire Protocol.

Required Methods

fn send_request(&mut self, piece: u32, offset: u32, length: u32) -> IoResult<()>

Sends a data request message to the remote peer.

fn send_block(&mut self, piece: u32, offset: u32, block: &[u8]) -> IoResult<()>

Sends a block ("piece") request to the remote peer.

fn send_cancel(&mut self, piece: u32, offset: u32, length: u32) -> IoResult<()>

Sends a cancel request to the remote peer.