Enum bittorrent::peer::PeerActionUnstable

pub enum PeerAction {
    Hidden,
    StateUpdate(StateChange),
    StateHave(u32),
    DataRequested(u32, u32, u32),
    DataArrived(u32, u32, Vec<u8>),
    DataCanceled(u32, u32, u32),
}

Denotes an action that was carried out in response to a message received by a remote peer.

Variants

Hidden

A message was received that has been hidden from the user (not guaranteed to propagate).

StateUpdate

A state message was received: (state_change)

StateHave

A have message was received: (piece_index)

DataRequested

A request message was received: (piece_index, block_offset, block_length)

DataArrived

A piece message was received: (piece_index, block_offset, block_data)

DataCanceled

A cancel message was received: (piece_index, block_offset, block_length)