Struct bittorrent::peer::handshake::HandshakeUnstable

pub struct Handshake {
    // some fields omitted
}

A struct representing a handshake that has successfully taken place.

Methods

impl Handshake

The entry point for connecting with remote peers.

fn initiate<T, S>(recipient: S, info: &InfoHash, curr_id: &PeerID, valid_peer_id: T) -> IoResult<Handshake> where T: for<'a> Fn<(&'a PeerID,), bool>, S: ToSocketAddr

Initiates a handshake with the recipient sending the designated info hash and peer id. If the response is malformed, the peer sends a different info hash, or the peer sends us a peer id that we are already using, the handshake will fail.

This is a blocking operation.

fn complete<T>(initiater: TcpStream, info: &InfoHash, curr_id: &PeerID, valid: T) -> IoResult<Handshake> where T: for<'a> Fn<(&'a PeerID,), bool>

Completes a handshake that was initiated by the remote peer. If the handshake initiated by the peer is malformed, the peer sent us a different info hash, or the peer sent us a peer id that we are already using, the handshake will fail.

fn into_peer(self, num_pieces: u32) -> Peer

Consumes the handshake object and creates a peer object with the number of pieces in the current torrent set to num_pieces.