Enum bittorrent::peer::message::PeerMessageUnstable

pub enum PeerMessage {
    Hidden,
    StateUpdate(StateChange),
    HaveUpdate(PieceIndex),
    BitfieldUpdate(Vec<u8>),
    BlockRequest(PieceIndex, BlockOffset, BlockLength),
    CancelRequest(PieceIndex, BlockOffset, BlockLength),
    BlockReceived(PieceIndex, BlockOffset, BlockLength),
    BlockReceivedTooBig(PieceIndex, BlockOffset, Vec<u8>),
}

Represents a message received from a remote peer.

Variants

Hidden

Message type has been hidden from client.

StateUpdate

Peer has changed the state for our end of the connection.

HaveUpdate

Peer has successfully downloaded a piece.

BitfieldUpdate

Peer has sent us a list of pieces they currently have.

BlockRequest

Peer has sent us a request for a block.

CancelRequest

Peer has sent us a cancel message for a block they requested previously.

BlockReceived

Peer has sent us a block of data that we requested.

BlockReceivedTooBig

Peer has sent us a block of data that we requested that is too big to fit in the buffer provided by the client.