Trait bittorrent::peer::message::PeerReaderUnstable

pub trait PeerReader {
    fn read_message<'a, F>(&mut self, max_pieces: u32, block: &mut F) -> IoResult<PeerMessage> where F: FnMut<(BlockLength,), &'a mut Block>;
    fn read_have(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>;
    fn read_bitfield(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>;
    fn read_request(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>;
    fn read_cancel(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>;
    fn read_block(&mut self, payload_len: u32, max_pieces: u32, block_buffer: &mut [u8]) -> IoResult<PeerMessage>;
}

Trait for reading Peer Wire Protocol messages.

Unlike the corresponding PeerWriter, the methods defined here should do bounds checking on both message lengths and piece indices in order to combat against buffer overflow attacks.

Required Methods

fn read_message<'a, F>(&mut self, max_pieces: u32, block: &mut F) -> IoResult<PeerMessage> where F: FnMut<(BlockLength,), &'a mut Block>

fn read_have(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>

fn read_bitfield(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>

fn read_request(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>

fn read_cancel(&mut self, payload_len: u32, max_pieces: u32) -> IoResult<PeerMessage>

fn read_block(&mut self, payload_len: u32, max_pieces: u32, block_buffer: &mut [u8]) -> IoResult<PeerMessage>