Module bittorrent::peer::stateUnstable

Structs

PeerState
Represents the state for one end of a connection between two peers.

Enums

StateChange
Represents a state change for an end of a connection in a peer to peer connection.

Constants

BITFIELD_ID

CHOKE_ID

HAVE_ID

HAVE_PAYLOAD_LEN

INTERESTED_ID

STATE_PAYLOAD_LEN

UNCHOKE_ID

UNINTERESTED_ID

Traits

StateSender
A trait dealing with state based messaging between a local and remote peer. This should be used as an interface for implementing the state messaging interface of the Bittorrent Peer Wire Protocol.