Trait bittorrent::peer::state::StateSenderUnstable

pub trait StateSender {
    fn send_state(&mut self, state: StateChange) -> IoResult<()>;
    fn send_have(&mut self, piece: u32) -> IoResult<()>;
    fn send_bitfield(&mut self, pieces: &[u8]) -> IoResult<()>;
}

A trait dealing with state based messaging between a local and remote peer. This should be used as an interface for implementing the state messaging interface of the Bittorrent Peer Wire Protocol.

Required Methods

fn send_state(&mut self, state: StateChange) -> IoResult<()>

Sends an update message to the peer where the message is determined by the state parameter passed in.

fn send_have(&mut self, piece: u32) -> IoResult<()>

Sends a have message to the peer where the piece is the index of the piece that we have successfully downloaded AND validated.

fn send_bitfield(&mut self, pieces: &[u8]) -> IoResult<()>

Sends a bitfield message to the peer where a 1 represents a piece that we have and a 0 represents a piece we are missing.

This MUST be sent immediately after the handshake and NOT any other time. We may choose to not send this if we have no pieces downloaded.