Struct bittorrent::peer::PeerUnstable

pub struct Peer {
    // some fields omitted
}

A Peer object representing a connection to a remote peer as well as the bittorrent state associated with that connection. Since the messaging interface between a peer is asynchronous, peers will start out as not choked and not interested.

Methods

impl Peer

fn local_state(&self) -> PeerState

Returns the state associated with our end of the connection.

fn remote_state(&self) -> PeerState

Returns the state associated with the remote end of the connection.

fn remote_peer_id(&self) -> &SPeerID

Returns a reference to a buffer containing the peer id of the remote peer.

fn process_messages<'a, T>(&mut self, messages: &mut [PeerMessage], block: &mut T) -> IoResult<()> where T: FnMut<(BlockLength,), &'a mut Block>

Processes all of the messages sent to us from the remote peer.

This method will not block while waiting for messages to be sent to us, but it will block if only part of a message was sent to us so far.

fn change_state(&mut self, state: StateChange) -> IoResult<()>

Sends a message to the remote peer telling them that we are changing their state.

fn notify_have(&mut self, piece: PieceIndex) -> IoResult<()>

Sends a message to the remote peer telling them that we have successfully downloaded and verified the hash of piece.

fn notify_bitfield(&mut self, bitfield: &[u8]) -> IoResult<()>

Sends a message to the remote peer telling them that the bits set in each byte correspond to the pieces we have downloaded.

To save bandwidth, partial bitfields combined with have messages to fill in spread out piece indices are generally allowed.

fn request_block(&mut self, piece: PieceIndex, offset: BlockOffset, len: BlockLength) -> IoResult<()>

Sends a message to the remote peer telling them that we are requesting a specific block of a specific length of a designated piece.

This is a data-oriented message and should not be sent if the peer is choking us (local end of connection has choked set to true).

fn cancel_block(&mut self, piece: PieceIndex, offset: BlockOffset, len: BlockLength) -> IoResult<()>

Sends a message to the remote peer telling them that we are cancelling a prior request for a block.

This is a data-oriented message and should not be sent if the peer is choking us (local end of connection has choked set to true).

fn send_block(&mut self, piece: PieceIndex, offset: BlockOffset, block_data: &[u8]) -> IoResult<()>

Sends a message to the remote peer telling them that we are sending a block of data to them.

This is a data-oriented message and should not be sent if the peer is choking us (local end of connection has choked set to true).