Struct bittorrent::torrent::TorrentUnstable

pub struct Torrent<'a> {
    // some fields omitted
}

A type representing a valid torrent file.

This type simply stores slices into other types that are actually holding the data. Therefore, it is very cheap to construct and copy around.

Methods

impl<'a> Torrent<'a>

fn new(metainfo: &'a Bencode) -> TorrResult<Torrent<'a>>

Goes through the Bencode object and pulls out the fields required by a valid torrent file. Any extraneous fields that may be present within the Bencode object are ignored and will not be considered an error.

fn announce(&self) -> &str

Returns the url of the main tracker for the current torrent file.

fn announce_list(&'a self) -> Option<&'a Vec<Vec<&str>>>

Optionally returns a list of urls pointing to backup trackers for the current torrent file.

fn comment(&'a self) -> Option<&'a str>

Optionally returns any comment within the current torrent file.

fn created_by(&'a self) -> Option<&'a str>

Optionally returns the created by tag within the current torrent file.

fn creation_date(&self) -> Option<i64>

Optionally returns the creation date of the current torrent file in standard UNIX epoch format.

Some RFCs say this should be a string, others say it should be an integer. In practice, all torrents I have seen use an integer so an error will be generated if it is anything other than an integer.

fn file_type(&self) -> (TorrFileType, &str)

Returns a tuple of the type of torrent file as well as the associated name. See below for what the second tuple value represents:

TorrFileType::Single -> str value will be the name of the file TorrFIleType::Multi -> str value will be the name of the root directory

fn files(&self) -> &Vec<TorrFile<'a>>

Returns a list of file objects for the current torrent.

If the torrent file is of type TorrFileType::Single, the list will have one entry. However, just because it has one entry does not necessarily mean it is of type TorrFileType::Single.

fn piece_length(&self) -> i64

Returns the piece length for each file.

fn pieces(&self) -> &[u8]

Returns the pieces byte array for the current torrent.