Module bittorrent::upnpUnstable

Structs

ServiceDesc
Type used to represent a service description. Service actions can be viewed as well as related in and out parameters and service state variables.

Enums

UPnPIntf
Type used to represent different interfaces available on a network as defined by the UPnP specification. This type can be used to get very general information about an interface but also to get access to the services exposed by the interface.

Type Definitions

StrPos
A tuple consisting of start and end position.