Enum bittorrent::upnp::UPnPIntfUnstable

pub enum UPnPIntf {
    // some variants omitted
}

Type used to represent different interfaces available on a network as defined by the UPnP specification. This type can be used to get very general information about an interface but also to get access to the services exposed by the interface.

Methods

impl UPnPIntf

fn find_all(from_addr: SocketAddr) -> IoResult<Vec<UPnPIntf>>

Sends out a search request for all UPnP interfaces on the specified from_addr.

This is a blocking operation.

fn find_services(from_addr: SocketAddr, name: &str, version: &str) -> IoResult<Vec<UPnPIntf>>

Sends out a search request for particular UPnP services on the specified from_addr.

This method sends out a search request asking that only services with the specified name and version respond. This method does not check to make sure that only services with the specified name and version responded. Additonally, any service interfaces that are not standard UPnP services http://upnp.org/sdcps-and-certification/standards/sdcps/ will not be found and should be discovered by using find_all and filtering the results.

This is a blocking operation.

fn find_devices(from_addr: SocketAddr, name: &str, version: &str) -> IoResult<Vec<UPnPIntf>>

Sends out a search request for particular UPnP devices on the specified from_addr.

This method sends out a search request asking that only devices with the specified name and version respond. This method does not check to make sure that only devices with the specified name and version responded. Additionally, any device interface that are not standard UPnP devices (http://upnp.org/sdcps-and-certification/standards/sdcps/) will not be found and should be discovered by using find_all and filtering the results.

This is a blocking operation.

fn find_uuid(from_addr: SocketAddr, uuid: &str) -> IoResult<Option<UPnPIntf>>

Sends out a search request for a unique UPnP interface on the specified from_addr.

This method sends out a search request asking that only the device with the specified uuid respond. This method does not check the the device that responded has the right uuid. If more than one device responds (which should not happen in any correct implementation) then the first device to respond will be returned.

This is a blocking operation.

fn is_root(&self) -> bool

Check if the current UPnPIntf represents a root device on the network.

fn is_device(&self) -> bool

Check if the current UPnPIntf represents a device on the network.

Note: This function will return false in the case of a root (even though it is also technically a device).

fn is_service(&self) -> bool

Check if the current UPnPIntf represents a service on the network.

fn is_identifier(&self) -> bool

Check if the current UPnPIntf represents an identifier (uuid) on the network.

Note: This function will return false in the case of a root or device even though all uuid's refer to a specific device (or root in some cases).

fn name<'a>(&'a self) -> &'a str

Get the device name, service specification, or uuid that this interface has identified itself with.

fn version<'a>(&'a self) -> &'a str

Get the device version, service version, or no version (in the case of an identifier or root) that this interface has identified itself with.

fn location<'a>(&'a self) -> &'a str

Get the location (url) of the root device description web page corresponding to this particular UPnPIntf.

fn st<'a>(&'a self) -> &'a str

Get the search target of this UPnPIntf.

fn usn<'a>(&'a self) -> &'a str

Get the full Unique Service Name of this UPnPIntf.

fn service_desc(&self) -> IoResult<ServiceDesc>

Get a service description object for this service.

Note: This will always fail if called on a non-service UPnPIntf.

This is a blocking operation.