Struct bittorrent::upnp::ServiceDescUnstable

pub struct ServiceDesc {
    // some fields omitted
}

Type used to represent a service description. Service actions can be viewed as well as related in and out parameters and service state variables.

Note: As of now, this object makes no guarantee that messages it sends to UPnP services that are not standard will be well formed; this is temporary as it allows the code to remain clean until a proper solution arises. A list of standardized UPnP devices and their corresponding services can be found at http://upnp.org/sdcps-and-certification/standards/sdcps/.

Methods

impl ServiceDesc

fn actions<'a>(&'a self) -> Vec<&'a str>

Returns a list of string slices corresponding to each action that can be sent to this service. The slice includes the action name, and any in or out parameters expected/provided as well as their related state variables. The "relatedStateVariable" corresponds to an entry in the "serviceStateTable".

The response is left in the XML format that it was sent from the UPnPIntf in. We leave it in this format so that we don't spend time allocating memory and parsing responses.

fn state_variables<'a>(&'a self) -> Option<&'a str>

Optionally returns a string slice corresponding to the "serviceStateTable" of the service description. This table explains the data types referenced in the "relatedStateVariable" field for each in or out parameters for each action.

The response is left in the XML format that it was sent from the UPnPIntf in. We leave it in this format so that we don't spend time allocating memory and parsing responses.

fn send_action(&self, action: &str, params: &[(&str, &str)]) -> IoResult<String>

Takes an action string representing the procedure to call as well as an array slice of tuples in the form (arg_name, arg_value). An empty array slice indicates that no parameters are expected for the specified action.

This operation will not check to make sure that valid actions or parameters have been passed in, so any error handling may happen on the service end.

This is a blocking operation.