Struct bittorrent::peer::Peer

pub struct Peer {
    // some fields omitted
}

A Peer object representing a connection to a remote peer as well as the bittorrent state associated with that connection. Since the messaging interface between a peer is asynchronous, peers will start out as not choked and not interested.

Methods

impl Peer

fn remote_peer_id(&self) -> &UPeerID

fn receive_messages(&mut self, actions: &mut [PeerAction]) -> IoResult<uint>